Share static options among classes
