+++ src/Carbon/CarbonInterface.php
@@ -2420,7 +2420,7 @@ interface CarbonInterface extends DateTimeInterface, JsonSerializable
      * Get the Carbon instance (real or mock) to be returned when a "now"
      * instance is created.
      *
-     * @return Closure|static the current instance used for testing
+     * @return Closure|CarbonInterface the current instance used for testing
      */
     public static function getTestNow();
