+++ src/Carbon/CarbonInterval.php
@@ -27,6 +27,7 @@ use Carbon\Traits\IntervalStep;
 use Carbon\Traits\MagicParameter;
 use Carbon\Traits\Mixin;
 use Carbon\Traits\Options;
+use Carbon\Traits\StaticOptionsLink;
 use Carbon\Traits\ToStringFormat;
 use Closure;
 use DateInterval;
@@ -197,6 +198,7 @@ class CarbonInterval extends DateInterval implements CarbonConverterInterface
         Mixin::mixin as baseMixin;
     }
     use Options;
+    use StaticOptionsLink;
     use ToStringFormat;
 
     /**
