+++ src/Carbon/Traits/Date.php
@@ -847,7 +847,6 @@ trait Date
     use Options;
     use Rounding;
     use Serialization;
-    use Test;
     use Timestamp;
     use Units;
     use Week;
