+++ src/Carbon/Traits/Options.php
@@ -29,125 +29,6 @@ trait Options
     use Localization;
 
     /**
-     * Days of weekend.
-     *
-     * @var array
-     */
-    protected static array $weekendDays = [
-        CarbonInterface::SATURDAY,
-        CarbonInterface::SUNDAY,
-    ];
-
-    /**
-     * Format regex patterns.
-     *
-     * @var array<string, string>
-     */
-    protected static array $regexFormats = [
-        'd' => '(3[01]|[12][0-9]|0[1-9])',
-        'D' => '(Sun|Mon|Tue|Wed|Thu|Fri|Sat)',
-        'j' => '([123][0-9]|[1-9])',
-        'l' => '([a-zA-Z]{2,})',
-        'N' => '([1-7])',
-        'S' => '(st|nd|rd|th)',
-        'w' => '([0-6])',
-        'z' => '(36[0-5]|3[0-5][0-9]|[12][0-9]{2}|[1-9]?[0-9])',
-        'W' => '(5[012]|[1-4][0-9]|0?[1-9])',
-        'F' => '([a-zA-Z]{2,})',
-        'm' => '(1[012]|0[1-9])',
-        'M' => '([a-zA-Z]{3})',
-        'n' => '(1[012]|[1-9])',
-        't' => '(2[89]|3[01])',
-        'L' => '(0|1)',
-        'o' => '([1-9][0-9]{0,4})',
-        'Y' => '([1-9]?[0-9]{4})',
-        'y' => '([0-9]{2})',
-        'a' => '(am|pm)',
-        'A' => '(AM|PM)',
-        'B' => '([0-9]{3})',
-        'g' => '(1[012]|[1-9])',
-        'G' => '(2[0-3]|1?[0-9])',
-        'h' => '(1[012]|0[1-9])',
-        'H' => '(2[0-3]|[01][0-9])',
-        'i' => '([0-5][0-9])',
-        's' => '([0-5][0-9])',
-        'u' => '([0-9]{1,6})',
-        'v' => '([0-9]{1,3})',
-        'e' => '([a-zA-Z]{1,5})|([a-zA-Z]*\\/[a-zA-Z]*)',
-        'I' => '(0|1)',
-        'O' => '([+-](1[012]|0[0-9])[0134][05])',
-        'P' => '([+-](1[012]|0[0-9]):[0134][05])',
-        'p' => '(Z|[+-](1[012]|0[0-9]):[0134][05])',
-        'T' => '([a-zA-Z]{1,5})',
-        'Z' => '(-?[1-5]?[0-9]{1,4})',
-        'U' => '([0-9]*)',
-
-        // The formats below are combinations of the above formats.
-        'c' => '(([1-9]?[0-9]{4})-(1[012]|0[1-9])-(3[01]|[12][0-9]|0[1-9])T(2[0-3]|[01][0-9]):([0-5][0-9]):([0-5][0-9])[+-](1[012]|0[0-9]):([0134][05]))', // Y-m-dTH:i:sP
-        'r' => '(([a-zA-Z]{3}), ([123][0-9]|0[1-9]) ([a-zA-Z]{3}) ([1-9]?[0-9]{4}) (2[0-3]|[01][0-9]):([0-5][0-9]):([0-5][0-9]) [+-](1[012]|0[0-9])([0134][05]))', // D, d M Y H:i:s O
-    ];
-
-    /**
-     * Format modifiers (such as available in createFromFormat) regex patterns.
-     *
-     * @var array
-     */
-    protected static array $regexFormatModifiers = [
-        '*' => '.+',
-        ' ' => '[   ]',
-        '#' => '[;:\\/.,()-]',
-        '?' => '([^a]|[a])',
-        '!' => '',
-        '|' => '',
-        '+' => '',
-    ];
-
-    /**
-     * Indicates if months should be calculated with overflow.
-     * Global setting.
-     *
-     * @var bool
-     */
-    protected static bool $monthsOverflow = true;
-
-    /**
-     * Indicates if years should be calculated with overflow.
-     * Global setting.
-     *
-     * @var bool
-     */
-    protected static bool $yearsOverflow = true;
-
-    /**
-     * Indicates if the strict mode is in use.
-     * Global setting.
-     *
-     * @var bool
-     */
-    protected static bool $strictModeEnabled = true;
-
-    /**
-     * Function to call instead of format.
-     *
-     * @var string|callable|null
-     */
-    protected static $formatFunction;
-
-    /**
-     * Function to call instead of createFromFormat.
-     *
-     * @var string|callable|null
-     */
-    protected static $createFromFormatFunction;
-
-    /**
-     * Function to call instead of parse.
-     *
-     * @var string|callable|null
-     */
-    protected static $parseFunction;
-
-    /**
      * Indicates if months should be calculated with overflow.
      * Specific setting.
      *
@@ -212,119 +93,6 @@ trait Options
      * @var string|callable|null
      */
     protected $localFormatFunction;
-
-    /**
-     * @deprecated To avoid conflict between different third-party libraries, static setters should not be used.
-     *             You should rather use the ->settings() method.
-     * @see settings
-     *
-     * Enable the strict mode (or disable with passing false).
-     *
-     * @param bool $strictModeEnabled
-     */
-    public static function useStrictMode(bool $strictModeEnabled = true): void
-    {
-        static::$strictModeEnabled = $strictModeEnabled;
-    }
-
-    /**
-     * Returns true if the strict mode is globally in use, false else.
-     * (It can be overridden in specific instances.)
-     *
-     * @return bool
-     */
-    public static function isStrictModeEnabled(): bool
-    {
-        return static::$strictModeEnabled;
-    }
-
-    /**
-     * @deprecated To avoid conflict between different third-party libraries, static setters should not be used.
-     *             You should rather use the ->settings() method.
-     *             Or you can use method variants: addMonthsWithOverflow/addMonthsNoOverflow, same variants
-     *             are available for quarters, years, decade, centuries, millennia (singular and plural forms).
-     * @see settings
-     *
-     * Indicates if months should be calculated with overflow.
-     *
-     * @param bool $monthsOverflow
-     *
-     * @return void
-     */
-    public static function useMonthsOverflow(bool $monthsOverflow = true): void
-    {
-        static::$monthsOverflow = $monthsOverflow;
-    }
-
-    /**
-     * @deprecated To avoid conflict between different third-party libraries, static setters should not be used.
-     *             You should rather use the ->settings() method.
-     *             Or you can use method variants: addMonthsWithOverflow/addMonthsNoOverflow, same variants
-     *             are available for quarters, years, decade, centuries, millennia (singular and plural forms).
-     * @see settings
-     *
-     * Reset the month overflow behavior.
-     *
-     * @return void
-     */
-    public static function resetMonthsOverflow(): void
-    {
-        static::$monthsOverflow = true;
-    }
-
-    /**
-     * Get the month overflow global behavior (can be overridden in specific instances).
-     *
-     * @return bool
-     */
-    public static function shouldOverflowMonths(): bool
-    {
-        return static::$monthsOverflow;
-    }
-
-    /**
-     * @deprecated To avoid conflict between different third-party libraries, static setters should not be used.
-     *             You should rather use the ->settings() method.
-     *             Or you can use method variants: addYearsWithOverflow/addYearsNoOverflow, same variants
-     *             are available for quarters, years, decade, centuries, millennia (singular and plural forms).
-     * @see settings
-     *
-     * Indicates if years should be calculated with overflow.
-     *
-     * @param bool $yearsOverflow
-     *
-     * @return void
-     */
-    public static function useYearsOverflow(bool $yearsOverflow = true): void
-    {
-        static::$yearsOverflow = $yearsOverflow;
-    }
-
-    /**
-     * @deprecated To avoid conflict between different third-party libraries, static setters should not be used.
-     *             You should rather use the ->settings() method.
-     *             Or you can use method variants: addYearsWithOverflow/addYearsNoOverflow, same variants
-     *             are available for quarters, years, decade, centuries, millennia (singular and plural forms).
-     * @see settings
-     *
-     * Reset the month overflow behavior.
-     *
-     * @return void
-     */
-    public static function resetYearsOverflow(): void
-    {
-        static::$yearsOverflow = true;
-    }
-
-    /**
-     * Get the month overflow global behavior (can be overridden in specific instances).
-     *
-     * @return bool
-     */
-    public static function shouldOverflowYears(): bool
-    {
-        return static::$yearsOverflow;
-    }
 
     /**
      * Set specific options.
